test/requirements.txt
jsonschema>=4.25.1,<5
PyYAML>=6.0,<7

